+++ eyes.common.java/src/main/java/com/applitools/connectivity/api/AsyncRequestCallback.java
@@ -1,0 +1,7 @@
+package com.applitools.connectivity.api;
+
+public interface AsyncRequestCallback {
+    void onComplete(Response response);
+
+    void onFail(Throwable throwable);
+}