+++ eyes.common.java/src/main/java/com/applitools/eyes/visualgrid/services/RenderingGridService.java
@@ -21,9 +21,12 @@ public class RenderingGridService extends Thread {
     private final Object concurrencyLock;
 
     public void setLogger(Logger logger) {
-
+        if (this.logger == null) {
+            this.logger = logger;
+        } else {
+            this.logger.setLogHandler(logger.getLogHandler());
+        }
     }
-
 
     public interface RGServiceListener {
         RenderingTask getNextTask();
@@ -43,6 +46,7 @@ public class RenderingGridService extends Thread {
     @Override
     public void run() {
         try {
+            logger.log("Service '" + this.getName() + "' had started");
             while (isServiceOn) {
                 if (isPaused) {
                     synchronized (debugLock) {
@@ -59,7 +63,7 @@ public class RenderingGridService extends Thread {
             if (this.executor != null) {
                 this.executor.shutdown();
             }
-            logger.verbose("Service '" + this.getName() + "' is finished");
+            logger.log("Service '" + this.getName() + "' is finished");
         } catch (Throwable e) {
             logger.verbose("Rendering Service Error : "+e);
         }