+++ eyes.sdk.core/src/main/java/com/applitools/connectivity/RemoteSessionEventHandler.java
@@ -1,56 +1,64 @@
-package com.applitools.eyes;
+package com.applitools.connectivity;
 
-import com.applitools.eyes.events.ISessionEventHandler;
+import com.applitools.connectivity.api.ConnectivityTarget;
+import com.applitools.connectivity.api.HttpClientImpl;
+import com.applitools.connectivity.api.Request;
+import com.applitools.connectivity.api.Response;
+import com.applitools.eyes.AbstractProxySettings;
+import com.applitools.eyes.Logger;
+import com.applitools.eyes.RectangleSize;
+import com.applitools.eyes.TestResults;
 import com.applitools.eyes.events.ValidationInfo;
 import com.applitools.eyes.events.ValidationResult;
+import com.applitools.utils.ArgumentGuard;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import javax.ws.rs.client.Entity;
-import javax.ws.rs.client.Invocation;
-import javax.ws.rs.client.WebTarget;
+import javax.ws.rs.HttpMethod;
 import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
 import java.net.URI;
 
-public class RemoteSessionEventHandler extends RestClient implements ISessionEventHandler {
+public class RemoteSessionEventHandler extends RestClient {
+
+    public static final int DEFAULT_CLIENT_TIMEOUT = 1000 * 60 * 5; // 5 minutes
+    private static final String SERVER_SUFFIX = "/applitools/sessions";
 
     private String autSessionId;
-
-    private static final String SERVER_SUFFIX = "/applitools/sessions";
     private final String accessKey;
-    private WebTarget defaultEndPoint;
+    private ConnectivityTarget defaultEndPoint;
     private boolean throwExceptions = true;
 
-    @SuppressWarnings("WeakerAccess")
     public RemoteSessionEventHandler(Logger logger, URI serverUrl, String accessKey, int timeout) {
         super(logger, serverUrl, timeout);
         this.accessKey = accessKey;
-        this.defaultEndPoint = endPoint.queryParam("accessKey", accessKey).path(SERVER_SUFFIX);
+        updateEndpoint();
     }
 
-    @SuppressWarnings("WeakerAccess")
     public RemoteSessionEventHandler(Logger logger, URI serverUrl, String accessKey) {
-        this(logger, serverUrl, accessKey, 30 * 1000);
-    }
-
-    public RemoteSessionEventHandler(URI serverUrl, String accessKey, int timeout) {
-        this(new Logger(), serverUrl, accessKey, timeout);
+        this(logger, serverUrl, accessKey, DEFAULT_CLIENT_TIMEOUT);
     }
 
     public RemoteSessionEventHandler(URI serverUrl, String accessKey) {
         this(new Logger(), serverUrl, accessKey);
     }
 
-    public void setProxy(AbstractProxySettings abstractProxySettings){
-        setProxyBase(abstractProxySettings);
-        this.defaultEndPoint = endPoint.queryParam("accessKey", accessKey).path(SERVER_SUFFIX);
+    private void updateEndpoint() {
+        this.defaultEndPoint = restClient.target(serverUrl).queryParam("accessKey", accessKey).path(SERVER_SUFFIX);
+    }
+
+    public void setProxy(AbstractProxySettings abstractProxySettings) {
+        super.setProxy(abstractProxySettings);
+        updateEndpoint();
+    }
+
+    public void setTimeout(int timeout) {
+        super.setTimeout(timeout);
+        updateEndpoint();
     }
 
     private void sendMessage(HttpMethodCall method) {
         StackTraceElement[] stackTraceElements = Thread.currentThread().getStackTrace();
         String methodName = "";
-        // getStackTrace()<-sendMessage()<-"actual caller"
         if (stackTraceElements.length >= 3) {
             methodName = stackTraceElements[2].getMethodName();
         }
@@ -58,8 +66,8 @@ public class RemoteSessionEventHandler extends RestClient implements ISessionEve
         Response response = null;
         try {
             response = method.call();
-            if (response.getStatus() != 200) {
-                logger.verbose("'" + methodName + "' notification handler returned an error: " + response.getStatusInfo());
+            if (response.getStatusCode() != 200) {
+                logger.verbose("'" + methodName + "' notification handler returned an error: " + response.getStatusPhrase());
             } else {
                 logger.verbose("'" + methodName + "' succeeded: " + response);
             }
@@ -75,76 +83,59 @@ public class RemoteSessionEventHandler extends RestClient implements ISessionEve
         }
     }
 
-    @Override
     public void initStarted() {
         sendMessage(new HttpMethodCall() {
             public Response call() {
-                Invocation.Builder invocationBuilder = defaultEndPoint.path(autSessionId)
-                        .request(MediaType.APPLICATION_JSON);
-
-                return invocationBuilder.put(Entity.json("{\"action\": \"initStart\"}"));
+                Request request = defaultEndPoint.path(autSessionId).request(MediaType.APPLICATION_JSON);
+                return request.method(HttpMethod.PUT, "{\"action\": \"initStart\"}", MediaType.APPLICATION_JSON);
             }
         });
     }
 
-    @Override
     public void initEnded() {
         sendMessage(new HttpMethodCall() {
             public Response call() {
-                Invocation.Builder invocationBuilder = defaultEndPoint.path(autSessionId)
-                        .request(MediaType.APPLICATION_JSON);
-
-                return invocationBuilder.put(Entity.json("{\"action\": \"initEnd\"}"));
+                Request request = defaultEndPoint.path(autSessionId).request(MediaType.APPLICATION_JSON);
+                return request.method(HttpMethod.PUT, "{\"action\": \"initEnd\"}", MediaType.APPLICATION_JSON);
             }
         });
     }
 
-    @Override
     public void setSizeWillStart(RectangleSize sizeToSet) {
         final RectangleSize size = sizeToSet;
         sendMessage(new HttpMethodCall() {
             public Response call() {
-                Invocation.Builder invocationBuilder = defaultEndPoint.path(autSessionId)
-                        .request(MediaType.APPLICATION_JSON);
-
-                return invocationBuilder.put(Entity.json("{\"action\": \"setSizeStart\", \"size\":{\"width\": " + size.getWidth() + ", \"height\": " + size.getHeight() + "}}"));
+                Request request = defaultEndPoint.path(autSessionId).request(MediaType.APPLICATION_JSON);
+                return request.method(HttpMethod.PUT, "{\"action\": \"setSizeStart\", \"size\":{\"width\": " + size.getWidth() + ", \"height\": " + size.getHeight() + "}}", MediaType.APPLICATION_JSON);
             }
         });
     }
 
-    @Override
     public void setSizeEnded() {
         sendMessage(new HttpMethodCall() {
             public Response call() {
-                Invocation.Builder invocationBuilder = defaultEndPoint.path(autSessionId)
-                        .request(MediaType.APPLICATION_JSON);
-
-                return invocationBuilder.put(Entity.json("{\"action\": \"setSizeEnd\"}"));
+                Request request = defaultEndPoint.path(autSessionId).request(MediaType.APPLICATION_JSON);
+                return request.method(HttpMethod.PUT, "{\"action\": \"setSizeEnd\"}", MediaType.APPLICATION_JSON);
             }
         });
     }
 
-    @Override
     public void testStarted(String autSessionId) {
         final String autSessionIdFinal = autSessionId;
         sendMessage(new HttpMethodCall() {
             public Response call() {
-                Invocation.Builder invocationBuilder = defaultEndPoint
-                        .request(MediaType.APPLICATION_JSON);
-
-                return invocationBuilder.post(Entity.json("{\"autSessionId\": \"" + autSessionIdFinal + "\"}"));
+                Request request = defaultEndPoint.request(MediaType.APPLICATION_JSON);
+                return request.method(HttpMethod.POST, "{\"autSessionId\": \"" + autSessionIdFinal + "\"}", MediaType.APPLICATION_JSON);
             }
         });
         this.autSessionId = autSessionId;
     }
 
-    @Override
     public void testEnded(String autSessionId, final TestResults testResults) {
         final String autSessionIdFinal = autSessionId;
         sendMessage(new HttpMethodCall() {
             public Response call() {
-                Invocation.Builder invocationBuilder = defaultEndPoint.path(autSessionIdFinal)
-                        .request(MediaType.APPLICATION_JSON);
+                Request invocationBuilder = defaultEndPoint.path(autSessionIdFinal).request(MediaType.APPLICATION_JSON);
                 // since the web API requires a root property for this message
                 ObjectMapper jsonMapper = new ObjectMapper();
                 String testResultJson;
@@ -155,34 +146,32 @@ public class RemoteSessionEventHandler extends RestClient implements ISessionEve
                     e.printStackTrace();
                 }
 
-                return invocationBuilder.put(Entity.json("{\"action\": \"testEnd\", \"testResults\":" + testResultJson + "}"));
+                return invocationBuilder.method(HttpMethod.PUT, "{\"action\": \"testEnd\", \"testResults\":" + testResultJson + "}", MediaType.APPLICATION_JSON);
             }
         });
     }
 
-    @Override
     public void validationWillStart(String autSessionId, final ValidationInfo validationInfo) {
         final String autSessionIdFinal = autSessionId;
         sendMessage(new HttpMethodCall() {
             public Response call() {
-                Invocation.Builder invocationBuilder = defaultEndPoint.path(autSessionIdFinal).path("validations")
+                Request request = defaultEndPoint.path(autSessionIdFinal).path("validations")
                         .request(MediaType.APPLICATION_JSON);
 
-                return invocationBuilder.post(Entity.json(validationInfo.toJsonString()));
+                return request.method(HttpMethod.POST, validationInfo.toJsonString(), MediaType.APPLICATION_JSON);
             }
         });
     }
 
-    @Override
     public void validationEnded(String autSessionId, final String validationId, final ValidationResult validationResult) {
         final String autSessionIdFinal = autSessionId;
         sendMessage(new HttpMethodCall() {
             public Response call() {
-                Invocation.Builder invocationBuilder = defaultEndPoint
+                Request request = defaultEndPoint
                         .path(autSessionIdFinal).path("validations").path(validationId)
                         .request(MediaType.APPLICATION_JSON);
 
-                return invocationBuilder.put(Entity.json("{\"action\":\"validationEnd\", \"asExpected\":" + validationResult.isAsExpected() + "}"));
+                return request.method(HttpMethod.PUT, "{\"action\":\"validationEnd\", \"asExpected\":" + validationResult.isAsExpected() + "}", MediaType.APPLICATION_JSON);
             }
         });
     }